+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 ## [Unreleased]
+### Added
+- Gene attributes with multiple values can be toggled to display more or less values
 ### Fixed
 - Adding Interproscan gff3 (```genenotebook add interproscan```)